pyproject.toml
[tool.poetry]
name = "cw"
version = "1.0.0"
description = ""
authors = ["Die Namen <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^1.4.36"
python-dotenv = "^0.20.0"
pydantic = "^1.8.0"
fastapi = "^0.75.2"
uvicorn = "^0.17.6"
asyncpg = "^0.25.0"
fastapi-pagination = "^0.9.3"
alembic = "^1.8.0"
psycopg2 = "^2.9.3"
sqlalchemy-pydantic-orm = "^0.1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
pytest-cov = "^3.0.0"
autoflake = "^1.4"
isort = "^5.10.0"
black = "^22.1"
unify = "^0.5"
flake8 = "^4.0"
pylint = "^2.12.1"
mypy = "^0.940"
pytest-deadfixtures = "^2.2.1"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
